% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{dist_mean}
\alias{dist_mean}
\title{Find mean of a probability distribution function}
\usage{
dist_mean(dist, l = 0, r = 1)
}
\arguments{
\item{dist}{A one-argument function}

\item{l}{Minimum value}

\item{r}{Maximum value}
}
\value{
A scalar
}
\description{
Find mean of a probability distribution function
}
\examples{

d1 <- update_prior(10, 40, P = 5/6, prior = stats::dunif)
dist_mean(d1)

}
