% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{empirical_bayes}
\alias{empirical_bayes}
\title{Estimate proportions of liars in multiple samples using empirical Bayes}
\usage{
empirical_bayes(heads, N, P)
}
\arguments{
\item{heads}{A vector of numbers of the good outcome reported}

\item{N}{A vector of sample sizes}

\item{P}{Probability of \emph{bad} outcome}
}
\value{
A list with two components:
\itemize{
\item \code{prior}, the calculated empirical prior (of class \code{densityFunction}).
\item \code{posterior}, a list of posterior distributions (objects of class \code{densityFunction}).
If \code{heads} was named, the list will have the same names.
}
}
\description{
This function creates a prior by fitting a Beta distribution to the \code{heads/N} vector,
using \code{\link[MASS:fitdistr]{MASS::fitdistr()}}. The prior is then updated using data from each
individual sample to give the posterior distributions.
}
\examples{

heads <- c(Baseline = 30, Treatment1 = 38, Treatment2 = 45)
N <- c(50, 52, 57)
res <- empirical_bayes(heads, N, P = 0.5)

compare_dists(res$posteriors$Baseline, res$posteriors$Treatment1)
plot(res$prior, ylim = c(0, 4), col = "grey", lty = 2)
plot(res$posteriors$Baseline, add = TRUE, col = "blue")
plot(res$posteriors$Treatment1, add = TRUE, col = "orange")
plot(res$posteriors$Treatment2, add = TRUE, col = "red")

}
