% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{print.densityFunction}
\alias{print.densityFunction}
\alias{plot.densityFunction}
\title{Print/plot an object of class \code{densityFunction}.}
\usage{
\method{print}{densityFunction}(x, ...)

\method{plot}{densityFunction}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused}
}
\description{
Print/plot an object of class \code{densityFunction}.
}
\examples{

d1 <- update_prior(33, 50, P = 0.5, prior = stats::dunif)
d1
plot(d1)

# show the actual R code (techies only)
unclass(d1)

}
