% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{update_prior}
\alias{update_prior}
\title{Calculate posterior of lambdas given prior, and heads reports of good
outcome}
\usage{
update_prior(heads, N, P, prior, npoints = 1000)
}
\arguments{
\item{heads}{Number of good outcomes reported}

\item{N}{Total number in sample}

\item{P}{Probability of \emph{bad} outcome}

\item{prior}{Prior over lambda. A function which takes a vector of values
between 0 and 1, and returns the probability density.}

\item{npoints}{How many points to integrate on?}
}
\value{
The probability density of the posterior distribution, as a one-argument
function.
}
\description{
Calculate posterior of lambdas given prior, and heads reports of good
outcome
}
\examples{
posterior <- update_prior(heads = 30, N = 50, P = 0.5, prior = stats::dunif)
plot(posterior)
}
