% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{compare_dists}
\alias{compare_dists}
\title{Estimate probability that two samples s1, s2 from 2 independent distributions have s1 > s2}
\usage{
compare_dists(dist1, dist2)
}
\arguments{
\item{dist1}{Density of distribution 1, as a one-argument function.}

\item{dist2}{Density of distribution 2.}
}
\value{
A probability scalar.
}
\description{
Estimate probability that two samples s1, s2 from 2 independent distributions have s1 > s2
}
\examples{

d1 <- update_prior(30, 50, P = 0.5, prior = stats::dunif)
d2 <- update_prior(25, 40, P = 0.5, prior = stats::dunif)
compare_dists(d1, d2)

}
