% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian-heads-cts.R
\name{difference_dist}
\alias{difference_dist}
\title{Given dist1 and dist2, find the pdf of dist1 - dist2}
\usage{
difference_dist(dist1, dist2)
}
\arguments{
\item{dist1, dist2}{Probability density functions}
}
\value{
A probability density function defined on \code{[-1, 1]}.
}
\description{
At the moment this only works when dist1 and dist2 are defined on \code{[0, 1]}.
}
\examples{

d1 <- update_prior(30, 50, P = 0.5, prior = stats::dunif)
d2 <- update_prior(25, 40, P = 0.5, prior = stats::dunif)
dd <- difference_dist(d1, d2)

}
